requirements.txt
# requirements.txt
streamlit
pyodbc
openai
python-dotenv
pandas